requirements/base.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt===5.3.1
gunicorn==23.0.0
kombu==5.4.2
packaging==24.1
prompt-toolkit==3.0.48
python-crontab==3.2.0
python-dateutil==2.9.0.post0 
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
psycopg2-binary